Fixed LEO address.
